Correction to version test
- The unknown version test can fail on certain of the random inputs
  because it is casting what is effectively a short to a byte, and
  the short will get truncated on large inputs. Two of these truncated
  results will be 0 and 1, which are the two numbers this test is
  expressly trying to avoid, because they are supposed to fail.
- The random range is now bound above by Byte.MAX_VALUE so that this
  truncating does not occur anymore.

+++ modP2p/test/org/aion/p2p/VerTest.java
@@ -43,7 +43,7 @@ public class VerTest {
         /*
          * inactive versions
          */
-        byte b1 = (byte) ThreadLocalRandom.current().nextInt(2, Short.MAX_VALUE);
+        byte b1 = (byte) ThreadLocalRandom.current().nextInt(2, Byte.MAX_VALUE);
         assertEquals(Ver.UNKNOWN, Ver.filter(b1));
     }
 }
